third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-kernelMatrix-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelMatrix' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('kernelMatrix') is "-2 0 0 0 1 0 0 0 2"
PASS convolveMatrixElement.getAttribute('kernelMatrix') is "3 0 3 0 0 0 3 0 3"
PASS successfullyParsed is true

TEST COMPLETE
